Fully proofread lesson (minor changes)
This commit reflects a full proofread through the lesson, where we fix various
minor issues such as typos, overlong sentences, inconsistent style, or instances
where the English doesn't sound natural.  Some of the changes are based on
personal preference, but I at least hope none are too controversial!

Note that we've opted for British English, including words ending '-ise' instead
of '-ize', and en-dashes (--) are used to separate clauses in a sentence.  The
majority of the document was in this style already, but should be more
consistent now.